修改数据类型

+++ app/src/main/java/com/zc741/xxx/ad/ClientAdNoticeNew.java
@@ -241,12 +241,12 @@ public class ClientAdNoticeNew {
              * content : 身体健康，万事如,意
              */
 
-            private int amount;
+            private float amount;
             private String headimgurl;
             private String salutation;
             private String content;
 
-            public int getAmount() {
+            public float getAmount() {
                 return amount;
             }
